English-center-student-management-BE/requirements.txt
Django==5.0.2
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
psycopg[binary]
python-decouple==3.8
drf-spectacular==0.27.0
pytest-django==4.7.0
pytest==7.4.4
pytest-cov==4.1.0
Pillow==11.0.0
django-cors-headers==4.3.1
django-filter==24.2
django-environ==0.11.2
gunicorn==21.2.0
